refactor: move ol search route to seperate file
